requirements.txt
Flask
Langchain
Pillow
requests
python-dotenv
torch
torchvision
numpy
opencv-python
flask-cors